+++ .storybook/config.js
@@ -1,5 +1,6 @@
 /**
  * @author Luciano Graziani @lgraziani2712
+ * @license {@link http://www.opensource.org/licenses/mit-license.php|MIT License}
  *
  */
 import 'babel-polyfill';