+++ src/components/pages/RawContentContainer.jsx
@@ -1,0 +1,14 @@
+/**
+ * @author Luciano Graziani @lgraziani2712
+ * @license {@link http://www.opensource.org/licenses/mit-license.php|MIT License}
+ *
+ * @flow
+ */
+import styled from 'styled-components';
+
+const RawContentContainer = styled.div`
+  margin: 0 auto;
+  width: 980px;
+`;
+
+export default RawContentContainer;